chore(status): Ensure that status is properly progogated in the /role path
