+++ client.go
@@ -67,6 +67,31 @@ func (e ejbcaAPIError) ToLogicalResponse() (*logical.Response, error) {
 	return response, err
 }
 
+func logicalResponseIsEjbcaError(resp *logical.Response) bool {
+	if resp == nil {
+		return false
+	}
+	if contentType, ok := resp.Data[logical.HTTPContentType].(string); !ok || contentType != "application/json" {
+		return false
+	}
+	if rawBody, ok := resp.Data[logical.HTTPRawBody].(string); !ok || rawBody == "" {
+		return false
+	}
+	if _, ok := resp.Data[logical.HTTPStatusCode].(int); !ok {
+		return false
+	}
+
+	err := json.Unmarshal([]byte(resp.Data[logical.HTTPRawBody].(string)), &resp.Data)
+	if err != nil {
+		return false
+	}
+	if _, ok := resp.Data["errors"]; !ok {
+		return false
+	}
+
+	return true
+}
+
 func (e *ejbcaClient) EjbcaAPIError(b *ejbcaBackend, detail string, err error) error {
 	logger := b.Logger().Named("ejbcaClient.createErrorFromEjbcaErr")
 	if err == nil {
